der: refactor traits into their own modules
This makes them easier to find, and also the monolithic `traits` module
was getting unwieldy and we're about ready to add some more.
