+++ der/src/encodable.rs
@@ -1,0 +1,60 @@
+//! Trait definition for [`Encodable`].
+
+use crate::{Encoder, Length, Result};
+
+#[cfg(feature = "alloc")]
+use {
+    crate::ErrorKind,
+    alloc::vec::Vec,
+    core::{convert::TryInto, iter},
+};
+
+/// Encoding trait.
+pub trait Encodable {
+    /// Compute the length of this value in bytes when encoded as ASN.1 DER.
+    fn encoded_len(&self) -> Result<Length>;
+
+    /// Encode this value as ASN.1 DER using the provided [`Encoder`].
+    fn encode(&self, encoder: &mut Encoder<'_>) -> Result<()>;
+
+    /// Encode this value to the provided byte slice, returning a sub-slice
+    /// containing the encoded message.
+    fn encode_to_slice<'a>(&self, buf: &'a mut [u8]) -> Result<&'a [u8]> {
+        let mut encoder = Encoder::new(buf);
+        self.encode(&mut encoder)?;
+        Ok(encoder.finish()?)
+    }
+
+    /// Encode this message as ASN.1 DER, appending it to the provided
+    /// byte vector.
+    #[cfg(feature = "alloc")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
+    fn encode_to_vec(&self, buf: &mut Vec<u8>) -> Result<Length> {
+        let expected_len = self.encoded_len()?.to_usize();
+        buf.reserve(expected_len);
+        buf.extend(iter::repeat(0).take(expected_len));
+
+        let mut encoder = Encoder::new(buf);
+        self.encode(&mut encoder)?;
+        let actual_len = encoder.finish()?.len();
+
+        if expected_len != actual_len {
+            return Err(ErrorKind::Underlength {
+                expected: expected_len.try_into()?,
+                actual: actual_len.try_into()?,
+            }
+            .into());
+        }
+
+        actual_len.try_into()
+    }
+
+    /// Serialize this message as a byte vector.
+    #[cfg(feature = "alloc")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
+    fn to_vec(&self) -> Result<Vec<u8>> {
+        let mut buf = Vec::new();
+        self.encode_to_vec(&mut buf)?;
+        Ok(buf)
+    }
+}
